Sync folders - erori identificate !! versiune inainte de corectare !!

+++ advanced_rsync/folder_sync.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import time
 
 def list_files(path):
     file_paths = []
@@ -58,6 +59,20 @@ def sync_folders(folder1, folder2):
     files1 = list_files(folder1)
     files2 = list_files(folder2)
 
+    # stergerea fisierelor care nu mai exista
+    for file in files1:
+        if file not in files2:
+            dst = os.path.join(folder1, file)
+            delete_file(dst)
+
+    for file in files2:
+        if file not in files1:
+            dst = os.path.join(folder2, file)
+            delete_file(dst)
+
+    files1 = list_files(folder1)
+    files2 = list_files(folder2)
+
     # sincronizare din folder1 in 2
     for file in files1:
         src = os.path.join(folder1, file)
@@ -65,8 +80,9 @@ def sync_folders(folder1, folder2):
 
         if file not in files2:
             copy_file(src, dst)
-        elif os.path.getmtime(src) > os.path.getmtime(dst):
-            copy_file(src, dst)
+        elif os.path.exists(dst) and os.path.exists(src):
+            if os.path.getmtime(src) > os.path.getmtime(dst):
+                copy_file(src, dst)
 
     # sincronizare din folder2 in 1
     for file in files2:
@@ -75,22 +91,12 @@ def sync_folders(folder1, folder2):
 
         if file not in files1:
             copy_file(src, dst)
-        elif os.path.getmtime(src) > os.path.getmtime(dst):
-            copy_file(src, dst)
-
-    # stergerea fisierelor care nu mai exista
-    for file in files1:
-        if file not in files2:
-            dst = os.path.join(folder2, file)
-            delete_file(dst)
-
-    for file in files2:
-        if file not in files1:
-            dst = os.path.join(folder1, file)
-            delete_file(dst)
-
+        elif os.path.exists(dst) and os.path.exists(src):
+            if os.path.getmtime(src) > os.path.getmtime(dst):
+                copy_file(src, dst)
 
 def run_sync(folder1, folder2):
     print(f"Starting continuous sync between:\n  {folder1}\n  {folder2}")
     while 1:
         sync_folders(folder1, folder2)
+        time.sleep(1)
